requirements.txt
PySide6==6.5.0
pillow==9.5.0
